Decription
Purpose of Document
This document is written for system developers: project managers, programmers, designers, system analysts, and testers. 

The purpose of writing this document, among others: 
a. Explain specifications of user requirements, scope, and limitations of the system built, processes, and target systems of the development team 
b. Equalizing perception between client and developer 
c. Become a description of the system to be built 
d. As documentation of information system requirements to be built 
e. Become a reference in system development